+++ epstein_files/documents/document.py
@@ -112,7 +112,7 @@ class Document:
     def __post_init__(self):
         self.file_info = FileInfo(self.file_path)
 
-        if not self.file_path.exists() and not self.file_info.is_local_extract_file and not self.file_info.is_doj_file:
+        if self.file_info.has_file and not self.file_path.exists():
             raise FileNotFoundError(f"File '{self.file_path}' does not exist!")
 
         self.text = self.text or self._load_file()