+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # NEXT RELEASE
+* Split up the massive Dilorio whistleblower emails
 
 ### 1.8.4